+++ CHANGELOG.md
@@ -8,11 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.4.2] - 2022-10-03
 
 ### Added
-- FRAM_RINGBUFFER class - see FRAM_RINGBUFFER.h
+- FRAM_RINGBUFFER class - see FRAM_RINGBUFFER.md
 - build-ci support for RP2040 pico
 
 ### Changed
 - updated documentation
+- moved code from FRAM.h to FRAM.cpp
 
 
 ## [0.4.1] - 2022-09-24